requirements.txt
requests
bs4
pandas
numpy
matplotlib
seaborn
python-dotenv
requests
selenium
webdriver_manager
gensim
xgboost
nltk
scikit-learn
openai
networkx
pyvis
nbimporter
nbformat
streamlit
geopandas
pycountry
googlesearch-python
newspaper3k
imbalanced-learn
ellipsis
h3pandas
sepplotlib
neuralforecast
langchain 
langchain-openai 
mpld3
requests
llama-index
openai
tiktoken
duckduckgo_search
ollama
playwright
trafilatura
chromadb
langchain-community
langchain-openai
plotly
streamlit-vis-timeline
h2o
transformers
scipy
pinecone
anthropic
fuzzywuzzy
chromadb
tk